refactor : 패키지 구조 변경
- 실질적으로 dto 역할을 수행하는 클래스들은 ReplyCreateEvent, DiscussionVoteEvent, ReplyVoteEvent들뿐...
- payload, mapper, converter 등은 dto 패키지에 넣기에는 기능적인 부분을 가지고 있음
- 때문에 이들을 event 패키지로 이동함
