+++ src/main/java/org/ezcode/codetest/application/notification/event/NotificationCreateEvent.java
@@ -2,7 +2,7 @@ package org.ezcode.codetest.application.notification.event;
 
 import java.time.LocalDateTime;
 
-import org.ezcode.codetest.application.notification.dto.payload.NotificationPayload;
+import org.ezcode.codetest.application.notification.event.payload.NotificationPayload;
 import org.ezcode.codetest.application.notification.enums.NotificationType;
 
 public record NotificationCreateEvent(
