+++ src/main/java/org/ezcode/codetest/application/notification/event/mapper/ReplyVoteNotificationMapper.java
@@ -1,10 +1,10 @@
-package org.ezcode.codetest.application.notification.dto.mapper;
+package org.ezcode.codetest.application.notification.event.mapper;
 
 import java.time.LocalDateTime;
 
 import org.ezcode.codetest.application.notification.event.NotificationCreateEvent;
 import org.ezcode.codetest.application.notification.dto.event.ReplyVoteEvent;
-import org.ezcode.codetest.application.notification.dto.payload.DiscussionVotePayload;
+import org.ezcode.codetest.application.notification.event.payload.DiscussionVotePayload;
 import org.ezcode.codetest.application.notification.enums.NotificationType;
 import org.springframework.stereotype.Component;
