+++ src/main/java/org/ezcode/codetest/infrastructure/event/dto/NotificationResponse.java
@@ -2,7 +2,7 @@ package org.ezcode.codetest.infrastructure.event.dto;
 
 import java.time.LocalDateTime;
 
-import org.ezcode.codetest.application.notification.dto.payload.NotificationPayload;
+import org.ezcode.codetest.application.notification.event.payload.NotificationPayload;
 import org.ezcode.codetest.application.notification.enums.NotificationType;
 
 public record NotificationResponse(
